move shutdown to system
